requirements.txt
opencv-python
onnxruntime
transformers
faiss-cpu
numpy
Pillow
fastapi
uvicorn
